requirements.txt
# Python requirements for restaurant menu generator
openpyxl
pandas
pillow
reportlab
python-telegram-bot